requirements.txt
# Core dependencies
neo4j==5.14.1
pandas==2.1.4
numpy==1.24.3
tqdm==4.66.1

# GraphRAG and LLM
langchain==0.1.0
openai==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# API and Web
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Data processing
pydantic==2.5.2
pyyaml==6.0.1

# Testing and development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
jupyter==1.0.0 